Move ecmult table computation code to separate file
