+++ src/precompute_ecmult.c
@@ -19,40 +19,7 @@
 #include "field_impl.h"
 #include "group_impl.h"
 #include "ecmult.h"
-
-/* Construct table of all odd multiples of gen in range 1..(2**(window_g-1)-1). */
-static void secp256k1_ecmult_compute_table(secp256k1_ge_storage* table, int window_g, const secp256k1_gej* gen) {
-    secp256k1_gej gj;
-    secp256k1_ge ge, dgen;
-    int j;
-
-    gj = *gen;
-    secp256k1_ge_set_gej_var(&ge, &gj);
-    secp256k1_ge_to_storage(&table[0], &ge);
-
-    secp256k1_gej_double_var(&gj, gen, NULL);
-    secp256k1_ge_set_gej_var(&dgen, &gj);
-
-    for (j = 1; j < ECMULT_TABLE_SIZE(window_g); ++j) {
-        secp256k1_gej_set_ge(&gj, &ge);
-        secp256k1_gej_add_ge_var(&gj, &gj, &dgen, NULL);
-        secp256k1_ge_set_gej_var(&ge, &gj);
-        secp256k1_ge_to_storage(&table[j], &ge);
-    }
-}
-
-/* Like secp256k1_ecmult_compute_table, but one for both gen and gen*2^128. */
-static void secp256k1_ecmult_compute_two_tables(secp256k1_ge_storage* table, secp256k1_ge_storage* table_128, int window_g, const secp256k1_ge* gen) {
-    secp256k1_gej gj;
-    int i;
-
-    secp256k1_gej_set_ge(&gj, gen);
-    secp256k1_ecmult_compute_table(table, window_g, &gj);
-    for (i = 0; i < 128; ++i) {
-        secp256k1_gej_double_var(&gj, &gj, NULL);
-    }
-    secp256k1_ecmult_compute_table(table_128, window_g, &gj);
-}
+#include "ecmult_compute_table_impl.h"
 
 static void print_table(FILE *fp, const char *name, int window_g, const secp256k1_ge_storage* table, int with_conditionals) {
     int j;
